changefeedccl: support a CSV format for changefeeds In this PR, we introduce a new CSV format for changefeeds. Note that this format is only supported with the initial_scan='only' option. For instance, one can now execute:
CREATE CHANGEFEED FOR foo WITH format=csv, initial_scan='only'

Release note (enterprise change): Support a CSV format for
changefeeds. Only works with initial_scan='only', and
does not work with diff/resolved options.
