+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -11,6 +11,9 @@ go_library(
         "changefeed_stmt.go",
         "doc.go",
         "encoder.go",
+        "encoder_avro.go",
+        "encoder_csv.go",
+        "encoder_json.go",
         "metrics.go",
         "name.go",
         "rowfetcher_cache.go",
@@ -87,6 +90,7 @@ go_library(
         "//pkg/util/ctxgroup",
         "//pkg/util/duration",
         "//pkg/util/encoding",
+        "//pkg/util/encoding/csv",
         "//pkg/util/envutil",
         "//pkg/util/hlc",
         "//pkg/util/httputil",
