+++ pkg/ccl/changefeedccl/testfeed_test.go
@@ -885,23 +885,34 @@ func (c *cloudFeed) Next() (*cdctest.TestFeedMessage, error) {
 			// The other TestFeed impls check both key and value here, but cloudFeeds
 			// don't have keys.
 			if len(m.Value) > 0 {
-				// Cloud storage sinks default the `WITH key_in_value` option so that
-				// the key is recoverable. Extract it out of the value (also removing it
-				// so the output matches the other sinks). Note that this assumes the
-				// format is json, this will have to be fixed once we add format=avro
-				// support to cloud storage.
-				//
-				// TODO(dan): Leave the key in the value if the TestFeed user
-				// specifically requested it.
-				var err error
-				if m.Key, m.Value, err = extractKeyFromJSONValue(m.Value); err != nil {
+				details, err := c.Details()
+				if err != nil {
 					return nil, err
 				}
-				if isNew := c.markSeen(m); !isNew {
-					continue
+
+				switch v := changefeedbase.FormatType(details.Opts[changefeedbase.OptFormat]); v {
+				case ``, changefeedbase.OptFormatJSON:
+					// Cloud storage sinks default the `WITH key_in_value` option so that
+					// the key is recoverable. Extract it out of the value (also removing it
+					// so the output matches the other sinks). Note that this assumes the
+					// format is json, this will have to be fixed once we add format=avro
+					// support to cloud storage.
+					//
+					// TODO(dan): Leave the key in the value if the TestFeed user
+					// specifically requested it.
+					if m.Key, m.Value, err = extractKeyFromJSONValue(m.Value); err != nil {
+						return nil, err
+					}
+					if isNew := c.markSeen(m); !isNew {
+						continue
+					}
+					m.Resolved = nil
+					return m, nil
+				case changefeedbase.OptFormatCSV:
+					return m, nil
+				default:
+					return nil, errors.Errorf(`unknown %s: %s`, changefeedbase.OptFormat, v)
 				}
-				m.Resolved = nil
-				return m, nil
 			}
 			m.Key, m.Value = nil, nil
 			return m, nil
@@ -1449,28 +1460,38 @@ func (f *webhookFeed) Next() (*cdctest.TestFeedMessage, error) {
 		if msg != "" {
 			m := &cdctest.TestFeedMessage{}
 			if msg != "" {
-				var err error
-				var resolved bool
-				resolved, err = isResolvedTimestamp([]byte(msg))
+				details, err := f.Details()
 				if err != nil {
 					return nil, err
 				}
-				if resolved {
-					m.Resolved = []byte(msg)
-				} else {
-					wrappedValue, err := extractValueFromJSONMessage([]byte(msg))
+				switch v := changefeedbase.FormatType(details.Opts[changefeedbase.OptFormat]); v {
+				case ``, changefeedbase.OptFormatJSON:
+					resolved, err := isResolvedTimestamp([]byte(msg))
 					if err != nil {
 						return nil, err
 					}
-					if m.Key, m.Value, err = extractKeyFromJSONValue(wrappedValue); err != nil {
-						return nil, err
+					if resolved {
+						m.Resolved = []byte(msg)
+					} else {
+						wrappedValue, err := extractValueFromJSONMessage([]byte(msg))
+						if err != nil {
+							return nil, err
+						}
+						if m.Key, m.Value, err = extractKeyFromJSONValue(wrappedValue); err != nil {
+							return nil, err
+						}
+						if m.Topic, m.Value, err = extractTopicFromJSONValue(m.Value); err != nil {
+							return nil, err
+						}
+						if isNew := f.markSeen(m); !isNew {
+							continue
+						}
 					}
-					if m.Topic, m.Value, err = extractTopicFromJSONValue(m.Value); err != nil {
-						return nil, err
-					}
-					if isNew := f.markSeen(m); !isNew {
-						continue
-					}
+				case changefeedbase.OptFormatCSV:
+					m.Value = []byte(msg)
+					return m, nil
+				default:
+					return nil, errors.Errorf(`unknown %s: %s`, changefeedbase.OptFormat, v)
 				}
 				return m, nil
 			}
@@ -1660,7 +1681,7 @@ func (p *pubsubFeed) Partitions() []string {
 
 // extractJSONMessagePubsub extracts the value, key, and topic from a pubsub message
 func extractJSONMessagePubsub(wrapped []byte) (value []byte, key []byte, topic string, err error) {
-	parsed := payload{}
+	parsed := jsonPayload{}
 	err = gojson.Unmarshal(wrapped, &parsed)
 	if err != nil {
 		return
@@ -1685,23 +1706,36 @@ func (p *pubsubFeed) Next() (*cdctest.TestFeedMessage, error) {
 	for {
 		msg := p.client.buffer.pop()
 		if msg != nil {
-			m := &cdctest.TestFeedMessage{}
-			resolved, err := isResolvedTimestamp([]byte(msg.data))
+			details, err := p.Details()
 			if err != nil {
 				return nil, err
 			}
-			msgBytes := []byte(msg.data)
-			if resolved {
-				m.Resolved = msgBytes
-			} else {
-				m.Value, m.Key, m.Topic, err = extractJSONMessagePubsub(msgBytes)
+
+			m := &cdctest.TestFeedMessage{}
+			switch v := changefeedbase.FormatType(details.Opts[changefeedbase.OptFormat]); v {
+			case ``, changefeedbase.OptFormatJSON:
+				resolved, err := isResolvedTimestamp([]byte(msg.data))
 				if err != nil {
 					return nil, err
 				}
-				if isNew := p.markSeen(m); !isNew {
-					continue
+				msgBytes := []byte(msg.data)
+				if resolved {
+					m.Resolved = msgBytes
+				} else {
+					m.Value, m.Key, m.Topic, err = extractJSONMessagePubsub(msgBytes)
+					if err != nil {
+						return nil, err
+					}
+					if isNew := p.markSeen(m); !isNew {
+						continue
+					}
 				}
+			case changefeedbase.OptFormatCSV:
+				m.Value = []byte(msg.data)
+			default:
+				return nil, errors.Errorf(`unknown %s: %s`, changefeedbase.OptFormat, v)
 			}
+
 			return m, nil
 		}
 		select {
